:lock: fix: Correção de problemas de segurança em rotas

+++ src/main/java/br/upe/devflix/controllers/VideoController.java
@@ -19,10 +19,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import br.upe.devflix.models.entities.Category;
+import br.upe.devflix.models.entities.User;
 import br.upe.devflix.models.entities.Video;
 import br.upe.devflix.services.CategoryCRUDService;
+import br.upe.devflix.services.UserCRUDService;
 import br.upe.devflix.services.VideoCRUDService;
 import br.upe.devflix.services.security.AuthorizationService;
+import br.upe.devflix.services.security.payload.JwtPayload;
 import br.upe.devflix.services.serializers.ResponseService;
 import br.upe.devflix.services.subsystems.YouTubeService;
 
@@ -35,6 +38,7 @@ public class VideoController {
   @Autowired private CategoryCRUDService categoryService;
   @Autowired private YouTubeService youtubeService;
   @Autowired private AuthorizationService authorizationService;
+  @Autowired private UserCRUDService userService;
 
   @GetMapping("/youtube/search")
   public ResponseEntity<?> searchByKeyword(
@@ -65,19 +69,31 @@ public class VideoController {
   {
     Category category = categoryService.fetch(categoryId);
     if (category == null){
+      //Categoria não encontrada...
       return responseService.create(null, HttpStatus.NOT_FOUND);
     }
     if (!authorizationService.isAuthenticated(authorization)){
+      //Usuário não está autenticado...
       return responseService.create(null, HttpStatus.FORBIDDEN);
     }
+
+    JwtPayload session = authorizationService.parseJwtPayload(authorization);
+    User owner = userService.fetch(session.getId());
+    if (owner.getId() != category.getOwner().getId()){
+      //Usuário não é o proprietário da categoria...
+      return responseService.create(null, HttpStatus.FORBIDDEN);
+    }
+
     List<Video> categoryVideos = category.getVideos();
-    Video addedVideo = videoService.insert(video.setCategory(category));
+    Video addedVideo = videoService.insert(
+      video.setCategory(category).setOwner(owner));
     categoryVideos.add(addedVideo);
     categoryService.update(category.setVideos(categoryVideos));
     return responseService.create(video, HttpStatus.OK);
   }
 
   public ResponseEntity<?> update(
+    @RequestHeader("authorization") String authorization,
     @PathVariable Long videoId,
     @RequestBody @Valid Video video)
   {
@@ -85,17 +101,40 @@ public class VideoController {
     if (foundVideo == null){
       return responseService.create(null, HttpStatus.NOT_FOUND);
     }
+    if (!authorizationService.isAuthenticated(authorization)){
+      //Usuário não está autenticado...
+      return responseService.create(null, HttpStatus.FORBIDDEN);
+    }
+
+    JwtPayload session = authorizationService.parseJwtPayload(authorization);
+    User owner = userService.fetch(session.getId());
+    if (owner.getId() != video.getOwner().getId()){
+      //Usuário não é o proprietário do vídeo...
+      return responseService.create(null, HttpStatus.FORBIDDEN);
+    }
     return responseService.create(
       videoService.update(videoId, video), HttpStatus.OK);
   }
 
   @DeleteMapping("/{videoId}")
   public ResponseEntity<?> delete(
+    @RequestHeader("authorization") String authorization,
     @PathVariable Long videoId)
   {
     Video foundVideo = videoService.fetch(videoId);
     if (foundVideo == null){
       return responseService.create(null, HttpStatus.NOT_FOUND);
+    }
+    if (!authorizationService.isAuthenticated(authorization)){
+      //Usuário não está autenticado...
+      return responseService.create(null, HttpStatus.FORBIDDEN);
+    }
+
+    JwtPayload session = authorizationService.parseJwtPayload(authorization);
+    User owner = userService.fetch(session.getId());
+    if (owner.getId() != foundVideo.getOwner().getId()){
+      //Usuário não é o proprietário do vídeo...
+      return responseService.create(null, HttpStatus.FORBIDDEN);
     }
     return responseService.create(
       videoService.delete(videoId), HttpStatus.OK);
